+++ HISTORY.md
@@ -3,7 +3,7 @@
 ## in development:
 * **New Features:**
     * pixels library: Added command blocks to rename and export costumes
-    * new "ide_all_blocks()" extension primitive, thanks, Michael!
+    * new "ide_blocks" extension primitive
     * new extension primitives: "txt_to_utf8(txt)" and "txt_from_utf8(utf8List)", thanks, John!
     * new extension primitive "txt_transform(name, txt)" for encoding, decoding, escaping, unescaping and hashing text
     * new extension primitive "cst_export(cst, name)" for exporting/downloading a costume
@@ -34,6 +34,7 @@
 * extensions: new "cst_export(cst, name)" extension primitive
 * pixels library: Added command blocks to rename and export costumes
 * extensions: fixed "Make Variables" extension for numerical names
+* extensions: fixed and renamed "ide_blocks" extension primitive
 
 ### 2022-09-21
 * extensions: refactored hyperized translation primitives